+++ medical_management/src/main/java/com/example/medical_management/model/account/AppRole.java
@@ -7,22 +7,22 @@ public class AppRole {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "role_id")
-    private int id;
+    private Long id;
     private String name;
 
     public AppRole() {
     }
 
-    public AppRole(int id, String name) {
+    public AppRole(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
